pyproject.toml
[tool.poetry]
name = "great-cards"
version = "0.1.0"
description = ""
authors = ["Grenui92 <grenui92@gmail.com>"]
readme = "README.md"
packages = [{include = "great_cards"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
django-environ = "^0.10.0"
openai = "^0.27.6"
psycopg2-binary = "^2.9.6"
pillow = "^9.5.0"
python-dotenv = "^1.0.0"
googletrans = "^3.0.0"
requests = "^2.30.0"
redis = "^4.5.5"
redis-lru = "^0.1.2"
django-cors-headers = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
